feature(GameObject) extract GameObject class
- extract GameObject class from App
- add a builder() method to GameObject

fix #9
